Migrating custom fields serializer
